Change double to BigDecimal

+++ src/main/java/com/u1s1/edq/entity/GeoVertex.java
@@ -1,9 +1,8 @@
 package com.u1s1.edq.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "GEO_VERTICES")
@@ -11,8 +10,8 @@ public class GeoVertex implements Serializable, Cloneable {
 
     private Integer id;
 
-    private double x_pos;
-    private double y_pos;
+    private BigDecimal x_pos;
+    private BigDecimal y_pos;
 
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Id
@@ -25,20 +24,20 @@ public class GeoVertex implements Serializable, Cloneable {
     }
 
     @Column(precision = 20, scale = 15)
-    public double getX_pos() {
+    public BigDecimal getX_pos() {
         return x_pos;
     }
 
-    public void setX_pos(double x_pos) {
+    public void setX_pos(BigDecimal x_pos) {
         this.x_pos = x_pos;
     }
 
     @Column(precision = 20, scale = 15)
-    public double getY_pos() {
+    public BigDecimal getY_pos() {
         return y_pos;
     }
 
-    public void setY_pos(double y_pos) {
+    public void setY_pos(BigDecimal y_pos) {
         this.y_pos = y_pos;
     }
